src/requirements.txt
# --- Dependencias principales de Flask ---
Flask==3.1.0
Flask-Cors==5.0.0
Werkzeug==3.1.3
Jinja2==3.1.4
itsdangerous==2.2.0
click==8.1.7
MarkupSafe==3.0.2
blinker==1.9.0

# --- Autenticación y seguridad ---
bcrypt==4.2.0
PyJWT==2.9.0

# --- Base de datos ---
PyMySQL==1.1.1

# --- Configuración y utilidades ---
python-dotenv==1.0.1
python-decouple==3.8
packaging==24.2
six==1.16.0
python-dateutil==2.9.0.post0
typing_extensions==4.12.2
Faker==33.0.0

# --- Producción ---
gunicorn==23.0.0
